python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-core-kit-python"
version = "0.1.0"
description = "Minimal AI provider and RAG examples in Python."
requires-python = ">=3.9"
dependencies = [
  "python-dotenv",
  "httpx",
  "pydantic"
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]